cmd: migrate page-item command to cobra style
