+++ cmd/bbolt/command_root.go
@@ -28,6 +28,7 @@ func NewRootCommand() *cobra.Command {
 		newPagesCommand(),
 		newKeysCommand(),
 		newDumpCommand(),
+		newPageItemCommand(),
 	)
 
 	return rootCmd
